Expose a `DateRangePicker.onCalendarKeyDown` prop that gets called when user presses a key while focus is inside the calendar ui.
For a11y reasons, we want to make sure that calendar is always closed when focus is outside the DateRangePicker component.
However, pressing Tab from the Calendar eventually sends focus outside of the DateRangePicker componene without closing the Calendar.
So, `onCalendarKeyDown` prop will act as a handler for the Tab key and all other unhandled key events