Pass the "event object" to the  "DateRangePicker.onClose" prop
React-dates closes the calendar, when user perform multiple actions
* user presses Shift+Tab from startDate field
* user presses Tab from endDate field
* user presses Enter from a calendar day
* user clicks on the of the DateRangePicker component
* user clicks outside of the DateRangePicker component
All there actions trigger the onClose callback, but do not provide it with the event object.
This makes it hard for calling component to make decision on how to procede once the calendar is closed
Example: move focus to desired element if user clicks on an non-focusable element outside of DateRangePicker component.
So, providing this event object will help in this desicion making process.